backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
PyMySQL==1.1.1
django-cors-headers==4.3.1
Pillow==11.1.0
python-decouple==3.8
